Made new function, game()
This function plays 5 rounds(playRound()), but alongside it keeps score for playerScore and computerScore.

After 5 rounds you got a text that you won, lost or tie. Result is based on playerScore and computerScore. You win if you have more points, lose if less, and tie if you have the same amount of points.